build: Bump version to 0.1.3
This commit updates the `shaderx` library version from `0.1.2` to `0.1.3`. The version bump is applied to the library configuration, the documentation, and the sample projects to ensure consistency.

### Key Changes:

- **Library Versioning:**
    - Updated the `version` property in `shaderx/build.gradle.kts` to `0.1.3`.
- **Documentation and Samples:**
    - Updated the implementation dependency in `README.md` to point to version `0.1.3`.
    - Updated the `shaderx` dependency version in the `VerticalCarousel` sample's `build.gradle.kts`.
