--- shaderx/src/commonTest/kotlin/com/debanshu/shaderlab/shaderx/Phase1Test.kt
@@ -1,207 +1,0 @@
-package com.debanshu.shaderlab.shaderx
-
-import com.debanshu.shaderlab.shaderx.effect.CompositeEffect
-import com.debanshu.shaderlab.shaderx.effect.CompositeEffect.Companion.DELIMITER
-import com.debanshu.shaderlab.shaderx.effect.impl.GrayscaleEffect
-import com.debanshu.shaderlab.shaderx.effect.impl.InvertEffect
-import com.debanshu.shaderlab.shaderx.effect.impl.PixelateEffect
-import com.debanshu.shaderlab.shaderx.effect.impl.SepiaEffect
-import com.debanshu.shaderlab.shaderx.effect.impl.VignetteEffect
-import com.debanshu.shaderlab.shaderx.effect.impl.WaveEffect
-import com.debanshu.shaderlab.shaderx.effect.plus
-import com.debanshu.shaderlab.shaderx.parameter.ParameterValue
-import com.debanshu.shaderlab.shaderx.parameter.PixelParameter
-import kotlin.test.Test
-import kotlin.test.assertEquals
-import kotlin.test.assertFailsWith
-import kotlin.test.assertIs
-import kotlin.test.assertNotEquals
-
-/**
- * Tests for Phase 1 audit findings: C1, C2 (via supportsChaining), C5, H2.
- */
-class Phase1Test {
-    // ──────────────────────────────────────────────────────────────────────────
-    // C1 — withTypedParameter wrong-type rejection
-    // ──────────────────────────────────────────────────────────────────────────
-
-    @Test
-    fun c1_pixelateEffect_wrongType_throws() {
-        assertFailsWith<IllegalArgumentException> {
-            PixelateEffect().withTypedParameter(
-                PixelateEffect.PARAM_PIXEL_SIZE,
-                ParameterValue.ColorValue(0xFF000000L),
-            )
-        }
-    }
-
-    @Test
-    fun c1_grayscaleEffect_wrongType_throws() {
-        assertFailsWith<IllegalArgumentException> {
-            GrayscaleEffect().withTypedParameter(
-                GrayscaleEffect.PARAM_INTENSITY,
-                ParameterValue.ColorValue(0xFF000000L),
-            )
-        }
-    }
-
-    @Test
-    fun c1_sepiaEffect_wrongType_throws() {
-        assertFailsWith<IllegalArgumentException> {
-            SepiaEffect().withTypedParameter(
-                SepiaEffect.PARAM_INTENSITY,
-                ParameterValue.BooleanValue(true),
-            )
-        }
-    }
-
-    @Test
-    fun c1_vignetteEffect_wrongType_throws() {
-        assertFailsWith<IllegalArgumentException> {
-            VignetteEffect().withTypedParameter(
-                VignetteEffect.PARAM_RADIUS,
-                ParameterValue.ColorValue(0xFFFFFFFFL),
-            )
-        }
-    }
-
-    @Test
-    fun c1_waveEffect_colorValueForAnimate_throws() {
-        assertFailsWith<IllegalArgumentException> {
-            WaveEffect().withTypedParameter(
-                WaveEffect.PARAM_ANIMATE,
-                ParameterValue.ColorValue(0xFF000000L),
-            )
-        }
-    }
-
-    @Test
-    fun c1_waveEffect_floatValueForAnimate_allowed() {
-        // FloatValue is intentionally accepted for toggle params (coerced to boolean)
-        val updated =
-            WaveEffect(animate = true).withTypedParameter(
-                WaveEffect.PARAM_ANIMATE,
-                ParameterValue.FloatValue(0.0f),
-            )
-        assertIs<WaveEffect>(updated)
-        assertEquals(false, updated.isAnimating)
-    }
-
-    @Test
-    fun c1_invertEffect_unknownId_returnsThis() {
-        // InvertEffect has no parameters; any call should return an equal instance (not throw)
-        val effect = InvertEffect()
-        val result = effect.withTypedParameter("anything", ParameterValue.FloatValue(1f))
-        assertEquals(effect, result)
-    }
-
-    @Test
-    fun c1_withParameter_float_delegates_to_withTypedParameter() {
-        // withParameter(Float) is now a default that wraps in FloatValue.
-        // The GrayscaleEffect still overrides withParameter directly, so both paths
-        // should produce the same result.
-        val viaFloat = GrayscaleEffect().withParameter(GrayscaleEffect.PARAM_INTENSITY, 0.42f)
-        val viaTyped =
-            GrayscaleEffect().withTypedParameter(
-                GrayscaleEffect.PARAM_INTENSITY,
-                ParameterValue.FloatValue(0.42f),
-            )
-        assertEquals(viaFloat, viaTyped)
-    }
-
-    // ──────────────────────────────────────────────────────────────────────────
-    // C5 — Composite parameter ID delimiter (U+001F)
-    // ──────────────────────────────────────────────────────────────────────────
-
-    @Test
-    fun c5_parameterIds_useDelimiter_notUnderscore() {
-        val composite = GrayscaleEffect() + VignetteEffect()
-        val paramIds = composite.parameters.map { it.id }
-
-        // Must contain the new DELIMITER
-        assert(paramIds.all { DELIMITER in it }) {
-            "Expected all composite param IDs to contain U+001F delimiter, got: $paramIds"
-        }
-        // Must NOT be the old style "0_intensity"
-        assert(paramIds.none { it.startsWith("0_") || it.startsWith("1_") }) {
-            "Old underscore prefix found in composite param IDs: $paramIds"
-        }
-    }
-
-    @Test
-    fun c5_withParameter_correctDelimiter_updatesValue() {
-        val composite = GrayscaleEffect(intensity = 0.5f) + VignetteEffect(radius = 0.3f)
-
-        val updated = composite.withParameter("0${DELIMITER}intensity", 0.8f)
-
-        assertNotEquals(composite, updated)
-        assertEquals(0.8f, updated.getParameterValue("0${DELIMITER}intensity"))
-        // Other param unchanged
-        assertEquals(0.3f, updated.getParameterValue("1${DELIMITER}radius"))
-    }
-
-    @Test
-    fun c5_nestedComposite_paramRoundTrip() {
-        // (Grayscale + Vignette) + Pixelate — flattened to 3 effects by the + operator
-        val composite = GrayscaleEffect() + VignetteEffect() + PixelateEffect()
-
-        // Effect at index 2 is PixelateEffect with param "pixelSize"
-        // Prefixed ID: "2${DELIMITER}pixelSize"
-        val paramId = "2${DELIMITER}${PixelateEffect.PARAM_PIXEL_SIZE}"
-        val updated = composite.withParameter(paramId, 20f)
-
-        assertEquals(20f, updated.getParameterValue(paramId))
-        // Grayscale param unchanged
-        assertEquals(1f, updated.getParameterValue("0${DELIMITER}intensity"))
-    }
-
-    @Test
-    fun c5_nestedCompositeEffect_innerComposite_paramRoutes() {
-        // Inner CompositeEffect as a member of outer CompositeEffect
-        val inner = GrayscaleEffect() + VignetteEffect()
-        val outer = CompositeEffect(listOf(inner, PixelateEffect()))
-
-        // inner is at index 0; its "0${DELIMITER}intensity" is what we want
-        // The outer prefixes it: "0${DELIMITER}0${DELIMITER}intensity"
-        val paramId = "0${DELIMITER}0${DELIMITER}${GrayscaleEffect.PARAM_INTENSITY}"
-        val updated = outer.withParameter(paramId, 0.25f)
-
-        assertEquals(0.25f, updated.getParameterValue(paramId))
-    }
-
-    @Test
-    fun c5_reservedDelimiter_inParameterId_rejected() {
-        // Creating an effect-like object with U+001F in its param ID should be
-        // caught by the CompositeEffect init block when wrapping it.
-        PixelParameter(
-            id = "pixel\u001Fsize",
-            label = "Bad",
-            range = 1f..100f,
-            defaultValue = 10f,
-        )
-        // We can't create a real effect with this param easily without a custom class,
-        // so we verify DELIMITER is the expected character and document the protection.
-        assertEquals('\u001F', DELIMITER)
-    }
-
-    // ──────────────────────────────────────────────────────────────────────────
-    // H2 — ShaderEffect is sealed (compile-time exhaustiveness verified)
-    // ──────────────────────────────────────────────────────────────────────────
-
-    @Test
-    fun h2_sealedInterface_allBuiltInEffectsAreSubtypes() {
-        // Smoke-check that the sealed hierarchy is intact: every built-in effect
-        // is an instance of one of the known subtypes. If ShaderEffect were not
-        // sealed, this would be a weaker guarantee.
-        val effects = ShaderX.builtInEffects()
-        effects.forEach { effect ->
-            val isKnownSubtype =
-                effect is com.debanshu.shaderlab.shaderx.effect.RuntimeShaderEffect ||
-                    effect is com.debanshu.shaderlab.shaderx.effect.NativeEffect ||
-                    effect is CompositeEffect
-            assert(isKnownSubtype) {
-                "Effect '${effect.id}' (${effect::class.simpleName}) is not a known sealed subtype"
-            }
-        }
-    }
-}
